Remove position calculation from cells

+++ Test/Test/Utils.h
@@ -8,8 +8,8 @@ class Utils
 {
 
 public:
-    static bool isPointEqual(int x1, int y1, int x2, int y2);
-    static bool isPointInSquare(int x, int y, int squareX1, int squareY1, int squareX2, int squareY2);
+    static bool isPointEqual(float x1, float y1, float x2, float y2);
+    static bool isPointInSquare(float x, float y, float squareX1, float squareY1, float squareX2, float squareY2);
     static int bitsetToInt(std::bitset<4> bitset);
     static int randomInt(int minNumber, int maxNumber);
